Rename libs/terraform to libs/deployplan
This is going to be working with both terraform and direct deployment
(https://github.com/databricks/cli/pull/2926).
