+++ bundle/deploy/terraform/plan.go
@@ -5,9 +5,9 @@ import (
 	"path/filepath"
 
 	"github.com/databricks/cli/bundle"
+	"github.com/databricks/cli/libs/deployplan"
 	"github.com/databricks/cli/libs/diag"
 	"github.com/databricks/cli/libs/log"
-	"github.com/databricks/cli/libs/terraform"
 	"github.com/hashicorp/terraform-exec/tfexec"
 )
 
@@ -51,7 +51,7 @@ func (p *plan) Apply(ctx context.Context, b *bundle.Bundle) diag.Diagnostics {
 	}
 
 	// Set plan in main bundle struct for downstream mutators
-	b.Plan = &terraform.Plan{
+	b.Plan = &deployplan.Plan{
 		Path:    planPath,
 		IsEmpty: !notEmpty,
 	}
